Fixed Local light Demo & Added FPS Counter
There are still bugs in the LocalLight script(s) which will cause
issues when debugging, but don't clearly effect a built version of the
game.

Scene 5 is the Local Light demo

A frames per second graphic appears in the top left of the screen. It
is found in the HUDFPS script file.

The HUDFPS script now attached to the player object